% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chromosomecoverageplot.R
\name{Chromosomecoverageplot}
\alias{Chromosomecoverageplot}
\title{Chromosomecoverageplot}
\usage{
Chromosomecoverageplot(filelistTable=df,gff=path to file,ylimvalue=c(-100,100),title="")
}
\arguments{
\item{filelistTable}{table of bedtools coverage output files with sample names as seperate column}

\item{gff}{path to gff file (parsed) , this file has to be parsed to include only gene features within the region of interest}

\item{ylimvalue}{y axis limit for the plots default c(-100,100)}
}
\value{
plot
}
\description{
function to plot read coverage for  region of intereset
}
\author{
pbaskaran
}
