% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gene_expressiontable.R
\name{create_gene_expressiontable}
\alias{create_gene_expressiontable}
\title{function to create exp table for different sample from same species}
\usage{
create_gene_expressiontable(path, pattern, header_number = 0,
  mergerby = "tracking_id", header = T, expression_column = 10,
  skip = 0, header_check = "yes")
}
\arguments{
\item{path}{to the common file folder}

\item{pattern}{\link{list.files} pattern to get input file name}

\item{header_number}{number of element to use as column header after spliting the full input filename}

\item{header}{booliean flag to indicate the expression file has header/not}

\item{expression_column}{index of expression value column}

\item{skip}{number of lines to skip before begining to read data}

\item{header_check}{checks the header in a looks "yes" dont check}

\item{mergeby}{merge different files using}
}
\value{
A matrix with expression value merged
}
\description{
function to create exp table for different sample from same species
}
\author{
Praveen Baskaran, \email{spemannpraveen@gmail.com}
}
