% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublePieOrDonutPlot.R
\name{donuts_plot}
\alias{donuts_plot}
\title{Function to create donut plot which is used to create double pie plot}
\usage{
donuts_plot(panel = runif(3), pctr = c(0.5, 0.2, 0.9), legend.label = "",
  cols = c("chartreuse", "chocolate", "deepskyblue"), outradius = 1,
  radius = 0.7, add = F, innerradius = 0.5, legend = F, pilabels = F,
  legend_offset = 0.25, borderlit = c(T, F, T, T), labelcex = 1,
  maintitle = "")
}
\arguments{
\item{pctr}{# percentage in count [Default = c(.5,.2,.9)]}

\item{legend.label}{legend label}

\item{cols}{colors}

\item{outradius}{outer radius [Default = 1]}

\item{radius}{1-width of the donus [Default =.7]}

\item{add}{Add dounut to the current plot [Default = F]}

\item{innerradius}{innerradius, if innerradius==innerradius then no suggest line [Default= .5]}

\item{legend}{display legend [Default = F]}

\item{pilabels}{display lables for each section of piece [Default = F]}

\item{legend_offset}{non-negative number, legend right position control [Default = .25]}

\item{borderlit}{vector of colors for pie border}

\item{labelcex}{size for lables}

\item{maintitle}{title for the plot}

\item{panle}{counts data}
}
\description{
Function to create donut plot

Run twice to create double layer pie plot
}
\examples{
### Then get the plot is fairly easy:
# INPUT data
browsers <- structure(list(browser = structure(c(3L, 3L, 3L, 3L, 2L, 2L,
                                                2L, 1L, 5L, 5L, 4L),
                                              .Label = c("Chrome", "Firefox", "MSIE","Opera", "Safari"),class = "factor"),
                          version = structure(c(5L,6L, 7L, 8L, 2L, 3L, 4L, 1L, 10L, 11L, 9L),
                                              .Label = c("Chrome 10.0", "Firefox 3.5", "Firefox 3.6", "Firefox 4.0", "MSIE 6.0",
                                                         "MSIE 7.0","MSIE 8.0", "MSIE 9.0", "Opera 11.x", "Safari 4.0", "Safari 5.0"),
                                             class = "factor"),
                          share = c(10.85, 7.35, 33.06, 2.81, 1.58,13.12, 5.43, 9.91, 1.42, 4.55, 1.65),
                          ymax = c(10.85, 18.2, 51.26,54.07, 55.65, 68.77, 74.2, 84.11, 85.53, 90.08, 91.73),
                          ymin = c(0,10.85, 18.2, 51.26, 54.07, 55.65, 68.77, 74.2, 84.11, 85.53,90.08)),
                     .Names = c("browser", "version", "share", "ymax", "ymin"),
                     row.names = c(NA, -11L), class = "data.frame")
## data clean
browsers=browsers[order(browsers$browser,browsers$share),]
arr=aggregate(share~browser,browsers,sum)
### choose your cols
mainCol=c("Blue","Green","Purple","Red","Grey")
donuts_plot(browsers$share,rep(1,11),browsers$version,
           cols=subcolors(browsers,"browser",mainCol),
          legend=F,pilabels = T,borderlit = rep(F,4) )
donuts_plot(arr$share,rep(1,5),arr$browser,
           cols=mainCol,pilabels=F,legend=T,legend_offset=-.02,
           outradius = .71,radius = .0,innerradius=.0,add=T,
           borderlit = rep(F,4) )
###end of line

}
\author{
Copied from internet and modified
}
