% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_functions.R
\name{mgsub}
\alias{mgsub}
\title{multiple gsub}
\usage{
mgsub(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{vector (mutliple/single) of pattern}

\item{replacement}{vector for replacement [should be same length as pattern]}

\item{x}{vector to search for}
}
\value{
mutlitple substituted x
}
\description{
multiple gsub function  :: multiple substitution
see \link{gsub} for more details
}
\examples{
 x<- c("a","b","a","c")
 mgsub(pattern=c("a","b"),replacement=c("x","y"),x=x)
}
\author{
praveen Baskran \email{spemannpraveen@gmail.com}
}
