% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc_functions.R
\name{multi_greplength}
\alias{multi_greplength}
\title{multi grep}
\usage{
multi_greplength(pattern, x, ...)
}
\arguments{
\item{pattern}{Multiple pattern in c()}

\item{x}{a vector}
}
\value{
number of matches for each pattern
}
\description{
grep vector with mutliple pattern and returns number of times each pattern match
\link{grep} This function uses grep.
}
\examples{
x<- c("a","b","a")
multi_greplength(pattern=c("a","b"),x=x)
}
\author{
praveen Baskran \email{spemannpraveen@gmail.com}
}
