% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_allsheets.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{read_excel_allsheets}
\usage{
read_excel_allsheets(file=xlx_filename)
}
\arguments{
\item{file}{Path to the excel filename}
}
\value{
returns a data frame with sheets info as comparison column. #### i.e (rbind of all sheeets)
}
\description{
Reads all sheets from a excel file

based on \url{http://stackoverflow.com/questions/12945687/how-to-read-all-worksheets-in-an-excel-workbook-into-an-r-list-with-data-frame-e}
}
\details{
function to plot read coverage for  region of intereset
}
\author{
praveen Baskran \email{spemannpraveen@gmail.com}
}
