% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublePieOrDonutPlot.R
\name{subcolors}
\alias{subcolors}
\title{function to create subclours for each donut
create}
\usage{
subcolors(.dta, main, mainCol)
}
\description{
function to create subclours for each donut
create
}
