% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write_Multifiles.R
\name{Write_Multifiles}
\alias{Write_Multifiles}
\title{Write multiple output file}
\usage{
Write_Multifiles(dTable, split_variable, outfolder = ".",
  append_text = "significant_differentially_expressed_genes.txt",
  html_output = FALSE, weblink_col = "gene_id", weblink = "",
  genExcel = T)
}
\arguments{
\item{dTable}{a dataframe}

\item{split_variable}{colname in dTable to split the table using using values in this colum}

\item{outfolder}{path to output folder otherwise "." or current working folder}

\item{append_text}{text to add to the file name [default= "significant_differentially_expressed_genes.txt"]}

\item{html_output}{save file also as sortable html table}

\item{genExcel}{a logical value. Generate a excel table addtionally. [Default= TRUE]}
}
\value{
write output files.
}
\description{
write multiple output file using split variable
see \link{ddply} for more details
}
\author{
praveen Baskran \email{spemannpraveen@gmail.com}
}
