% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeProbeDataHTML.R
\name{writeProbeDataHTML}
\alias{writeProbeDataHTML}
\title{Write Html output file for pfam enrichment file}
\usage{
writeProbeDataHTML(x, filename, col.format, id.header = "id",
  col.names = NA, title = "Gene List", header = NA, text = NA,
  indices = TRUE, datestamp = TRUE, css = TRUE, js = TRUE, pad = "",
  weblink_col = "gene_id", weblink = "")
}
\value{
Save files in  html files
}
\description{
write sortable html file using Pfam enrichment table

This is a modified version of "writeProbeData" function from "probeAnnotation" package
see \link{writeProbeData} for more details and for options
}
\author{
praveen Baskran \email{spemannpraveen@gmail.com}
}
